moved old classes to GraphModelLibrary.Legacy namespace
